Skeleton Outlin
We are currently setting up the skeletal outline of the webpages with
the expectation that we’ll be uploading images and more appropriate
descriptions.

A flow chart has also been included to make it easier to navigate and
finalize the website. As of this moment we are expecting 42 pages for
this site (40 shown on the chart plus homepage.html and about.html)