Creation of page13.html
This webpage is different from the others in that we have four boxes to
deal with instead of two. The original css file was written to handle
two boxes in a webpage but this complicated it briefly until I found a
code sequence that let four boxes to be created, evenly spaced out.